make AbstractFileStringIntStream and AbstractFileStringStream inner classes of StringsTextExtractor
